+++ tests/upgrade_http.rs
@@ -9,7 +9,7 @@ use std::net::SocketAddr;
 use anyhow::Result;
 use axum::{routing, Router};
 use axum_server_dual_protocol::{ServerExt, UpgradeHttpLayer};
-use http::header::LOCATION;
+use http_0::header::LOCATION;
 use reqwest::redirect::Policy;
 use reqwest::{Certificate, Client, StatusCode};